Implement base utility classes and types
A client (pinbot) I have made to use Artieficer as a framework had
implemented various functionalities such as it's own ExtendedClient and
Configuration type, which every single one of my future bots could
potentially benefit from, so they were implemented here to be used as a
sort of "standard".

Some minor refactoring was done, such as adding a validation for a new
configuration file property and changing the log type from error to
warning for when a bot was disabled.

The README at src/clients/root was updated to reflect the new
configuration property, although this documentation will be moved to a
future wiki for the repository.

Bumped version from v0.1.2 to v0.1.3